+++ src/runtime/components/Input.vue
@@ -168,7 +168,7 @@ function updateInput(value: string | null | undefined) {
     value ||= null
   }
 
-  if (props.modelModifiers?.optional) {
+  if (props.modelModifiers?.optional && !props.modelModifiers?.nullable && value !== null) {
     value ||= undefined
   }